requirements.txt
Django==1.3.1
-e git://github.com/clincher/ecigar-django-shop.git#egg=django-shop
pillow
django-simple-captcha
South==0.7.3
django-classy-tags==0.3.4.1
django-debug-toolbar==0.8.5
django-feedback==0.4
django-filebrowser==3.3.0
django-grappelli==2.3.4
django-polymorphic==0.2
django-ratings==0.3.6
django-shop-simplevariations
django-sitetree==0.4
django-tinymce==1.5.1a2
djangocms-utils==0.9.5
ipython==0.11
sorl-thumbnail==11.05.2
wsgiref==0.1.2
